+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/experimental/DataStreamUtils.java
@@ -15,8 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.util.datastream;
+package org.apache.flink.streaming.experimental;
 
+import org.apache.flink.annotation.PublicEvolving;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.runtime.net.ConnectionUtils;
 import org.apache.flink.streaming.api.datastream.DataStream;
@@ -33,7 +34,11 @@ import java.util.Iterator;
 
 /**
  * A collection of utilities for {@link DataStream DataStreams}.
+ *
+ * <p>This experimental class is relocated from flink-streaming-contrib. Please see package-info.java
+ * for more information.
  */
+@PublicEvolving
 public final class DataStreamUtils {
 
 	/**
