+++ flink-streaming-scala/src/main/scala/org/apache/flink/streaming/experimental/package.scala
@@ -16,11 +16,11 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.utils
+package org.apache.flink.streaming.experimental
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
-import org.apache.flink.streaming.util.datastream.{DataStreamUtils => JavaStreamUtils}
 import org.apache.flink.streaming.api.scala.DataStream
+import org.apache.flink.streaming.experimental.{DataStreamUtils => JavaStreamUtils}
 
 import scala.reflect.ClassTag
 import scala.collection.JavaConverters._
@@ -29,7 +29,9 @@ package object utils {
 
   /**
     * This class provides simple utility methods for collecting a [[DataStream]],
-    * effectively enriching it with the functionality encapsulated by [[JavaStreamUtils]].
+    * effectively enriching it with the functionality encapsulated by [[DataStreamUtils]].
+    *
+    * This experimental class is relocated from flink-streaming-contrib.
     *
     * @param self DataStream
     */
